Finish part 1/4
Box Model
Because everything CSS displays is a box, understanding how the CSS Box Model works is a core foundation of CSS.

Selectors
To apply CSS to an element, you need to select it. CSS gives you a number of different ways to do this, and you can explore them in this module.

The cascade
Sometimes two or more competing CSS rules can apply to an element. Find out how the browser chooses which to use, and how to control this selection.

Specificity
This module takes a deeper look at specificity, a key part of the cascade.

Inheritance
Some CSS properties inherit if you don't specify a value for them. Find out how this works and how to use it to your advantage in this module.

Color
There are several different ways to specify color in CSS. This module examines the most commonly used color values.